+++ src/AsteroidGroups/Flyby.cs
@@ -51,7 +51,7 @@ namespace Starstrider42.CustomAsteroids {
 
 			this.detectable = null;
 
-			this.classRatios = null;
+			this.classRatios = new Proportions<string>(new [] {"1.0 PotatoRoid"});
 		}
 
 		public double getSpawnRate() {
@@ -286,15 +286,14 @@ namespace Starstrider42.CustomAsteroids {
 			return 0.5 * (innerUT + outerUT);
 		}
 
-		public ConfigNode drawAsteroidData() {
+		public string drawAsteroidType() {
 			try {
-				AsteroidType typeInfo = AsteroidManager.drawAsteroidType(classRatios);
-				return typeInfo.packedAsteroidData();
+				return AsteroidManager.drawAsteroidType(classRatios);
 			} catch (InvalidOperationException e) {
 				Debug.LogWarning("[CustomAsteroids]: Could not select asteroid class; reverting to default.");
 				Debug.LogException(e);
 
-				return new CustomAsteroidData().toProtoConfigNode();
+				return "PotatoRoid";
 			}
 		}
 	}