+++ src/AsteroidGroups/Population.cs
@@ -65,7 +65,7 @@ namespace Starstrider42.CustomAsteroids {
 
 			this.detectable = null;
 
-			this.classRatios = null;
+			this.classRatios = new Proportions<string>(new [] {"1.0 PotatoRoid"});
 		}
 
 		public Orbit drawOrbit() {
@@ -178,15 +178,14 @@ namespace Starstrider42.CustomAsteroids {
 			}
 		}
 
-		public ConfigNode drawAsteroidData() {
+		public string drawAsteroidType() {
 			try {
-				AsteroidType typeInfo = AsteroidManager.drawAsteroidType(classRatios);
-				return typeInfo.packedAsteroidData();
+				return AsteroidManager.drawAsteroidType(classRatios);
 			} catch (InvalidOperationException e) {
 				Debug.LogWarning("[CustomAsteroids]: Could not select asteroid class; reverting to default.");
 				Debug.LogException(e);
 
-				return new CustomAsteroidData().toProtoConfigNode();
+				return "PotatoRoid";
 			}
 		}