+++ CHANGELOG.md
@@ -4,11 +4,12 @@ Version History                         {#changelog}
 Custom Asteroids conforms to version 2.0.0 of the [Semantic Versioning specification](http://semver.org/spec/v2.0.0.html). 
 All version numbers are to be interpreted as described therein. In addition to the actual [public API](http://starstrider42.github.io/Custom-Asteroids/), the [format of asteroid population files](http://starstrider42.github.io/Custom-Asteroids/newbelts.html) and the information stored in KSP save files will be considered part of the API for the purpose of versioning.
 
-Version 1.3.0 (release pending)
+Version 1.3.0
 ------------
 
 ### New Features
 
+* Support for KSP 1.1.
 * Limited support for custom asteroid types.
 * Can now customize asteroids on intercept trajectories.
 * Asteroids can now appear only under certain conditions.